refactoring project struct
